fix: static mutable state in marshalTables
 - port to .js
